Move ContentRangeParser to datasources.utils package
Developed using AI-assisted tooling
